Self-tests: add logic for round of tests verifying previous MPI tests

+++ tests/test_mpi_wrapper.sh
@@ -2,7 +2,9 @@
 
 test_name=$1
 test_dir=$(pwd)
-sandbox="$test_name""_sandbox"
+sandbox="$test_name""_sandbox/"
+prereq_test_name=${3:-}
+prereq_sandbox="$prereq_test_name""_sandbox/"
 args=$2
 configuration_file=SciATHBatchQueuingSystem.conf
 python=python
@@ -12,8 +14,12 @@ if [ ! -f "$configuration_file" ]; then
   exit 1
 fi
 
-rm -rf $sandbox
-mkdir $sandbox
+if [ -z "$prereq_test_name" ]; then
+  rm -rf $sandbox
+  mkdir $sandbox
+else
+  cp -R $prereq_sandbox $sandbox
+fi
 cp $configuration_file $sandbox
 cd $sandbox
 $python -m sciath $args \
